mariadDB4j-maven-plugin goal start, stop
This is used to bring up a database before an webapplication or simliar is booted.
Where said application does not have easy hooks to start an embedded database easily
and is only used to emulate higher environment without need to deal with differinces
in database language or configuration options.

Can be used in combination with build-helper:reserve-network-port and spring-boot:start,stop
ensure that reserve-network-port is before this plugin and spring-boot is after, else
the spring application will be started before the database (maven xml ordering matters)