Issue #259
Add new signatures for source and run methods which include booelan force argument which, if true, add the -f (--force) argument to the mysql client process. This means mysql will continue with processing if it encounters an error in a statement. The protected run method does the work of adding the argument. Existing source() and run() methods have been modifed to call the new method passing false for the force argument to maintain current functionality.
Added self to CONTRIBUTORS.md