Fixes #257 - maven-checkstyle-plugin:3.1.0
The LeftCurlyCheck maxLineLength property has been removed. The property allowed for the check to be excepted for lines longer than the specified value and the left curly bracket could be on a new line. If this is happening, then you have bigger problems than an extra two chars on the end of your line, hence it was removed. I have simply removed the property as it should not be required, and indeed IS NOT required in MariaDB4j. There are no checkstyle errors following its removal.

I have only commited the required checkstyle.xml change. I am assuming dependabot will make the necessary pom.xml changes when #257 is rebased?